This is home and end position all function is OK,by test batter,score 3.0, DATA-0202-0301

+++ hardware/bldc/bldc.h
@@ -60,7 +60,7 @@ extern __IO int32_t  uwStep ;
 
 extern __IO uint32_t Lock_Time; // 堵转时间
 extern __IO uint16_t   PWM_Duty;   //占空比
-extern __IO uint16_t  PID_PWM_Duty ;
+extern __IO int32_t  PID_PWM_Duty ;
 
 extern __IO int8_t Dir;
